split out some v1 specific consts
