+++ apps/client-api/routes/v1/comments.js
@@ -1,11 +1,12 @@
 const { Op } = require('sequelize');
 const { Router } = require('express');
 const { fixchar } = require('fixchar');
-const { RESPONSE_FIELDS, CACHE_TTL } = require('../../../../consts');
+const { CACHE_TTL } = require('../../../../consts');
 const { db, GenericError } = require('../../../../modules');
 const { asyncMiddleware } = require('../../middleware/error');
 const { limitChecker } = require('../../middleware/filters');
 const cacheService = require('../../services/CacheService');
+const { RESPONSE_FIELDS } = require('../../../../consts/v1_consts');
 
 const router = new Router();
