+++ apps/client-api/routes/v1/videos.js
@@ -2,9 +2,10 @@ const moment = require('moment-timezone');
 const { Op } = require('sequelize');
 const { Router } = require('express');
 const { db } = require('../../../../modules');
-const { RESPONSE_FIELDS, ORGANIZATIONS } = require('../../../../consts');
+const { ORGANIZATIONS } = require('../../../../consts');
 const { asyncMiddleware } = require('../../middleware/error');
 const { limitChecker } = require('../../middleware/filters');
+const { RESPONSE_FIELDS } = require('../../../../consts/v1_consts');
 
 const router = new Router();
